petites librairies de c++
Some codes I got to used in a financial purpose mainly but not only.
The aim of the functions is to gives a result thus the implementation is
not optimal and there are many ways to speed the compilation up, I agree
and accept it but it wasn't my main concern at the time of
implementation.